implemented currency_breakdown function

+++ basic_functions.c
@@ -78,7 +78,7 @@ int greatest_of_three(int num1, int num2, int num3)
 
 float average_of_three(int num1, int num2, int num3)
 {
-  return (num1 + num2 + num3) / 3;
+  return (num1 + num2 + num3) / (float)3;
 }
 
 int main(void)
